requirements.txt
pyyaml==5.4.1
docxtpl==0.18.0